Refactor out ToolButton
